+++ inst/example/simulations.R
@@ -21,14 +21,19 @@ mxl_pref <- readRDS(here::here('inst', 'extdata', 'mxl_pref.Rds'))
 mxl_wtp  <- readRDS(here::here('inst', 'extdata', 'mxl_wtp.Rds'))
 
 # Create a set of alternatives for which to simulate shares. Each row is an
-# alternative and each column an attribute. In this example, I just use one
-# of the choice observations from the yogurt dataset:
-alts <- subset(yogurt, obsID == 42, select = c('price', 'feat', 'brand'))
-row.names(alts) <- c('dannon', 'hiland', 'weight', 'yoplait')
+# alternative and each column an attribute. In this example, I just use a
+# couple of the choice observations from the yogurt dataset:
+alts <- subset(yogurt, obsID %in% c(42, 13),
+               select = c('price', 'feat', 'brand'))
+alts$obsID <- c(rep(1, 4), rep(2, 4))
 alts
 
 # Run the simulation using the preference space MNL model:
-sim_mnl_pref <- simulateShares(mnl_pref, alts, alpha = 0.025)
+sim_mnl_pref <- simulateShares(
+  model     = mnl_pref,
+  alts      = alts,
+  obsIDName = "obsID"
+)
 sim_mnl_pref
 
 # The results show the expected shares for each alternative.
@@ -37,7 +42,12 @@ sim_mnl_pref
 # value (e.g. a 90% CI is obtained with alpha = 0.05).
 
 # Run the simulation using the WTP space MNL model:
-sim_mnl_wtp <- simulateShares(mnl_wtp, alts, priceName = 'price')
+sim_mnl_wtp <- simulateShares(
+  model     = mnl_wtp,
+  alts      = alts,
+  obsIDName = "obsID",
+  priceName = 'price'
+)
 sim_mnl_wtp
 
 # Since these two models are equivalent except in different spaces, the
@@ -46,10 +56,19 @@ sim_mnl_wtp
 # WTP space models.
 
 # Simulations can also be run using MXL models in either space:
-sim_mxl_pref <- simulateShares(mxl_pref, alts)
+sim_mxl_pref <- simulateShares(
+  model     = mxl_pref,
+  alts      = alts,
+  obsIDName = "obsID"
+)
 sim_mxl_pref
 
-sim_mxl_wtp <- simulateShares(mxl_wtp, alts, priceName = 'price')
+sim_mxl_wtp <- simulateShares(
+  model     = mxl_wtp,
+  alts      = alts,
+  obsIDName = "obsID",
+  priceName = 'price'
+)
 sim_mxl_wtp
 
 # Save results
@@ -66,13 +85,14 @@ saveRDS(sim_mxl_wtp,
 library(ggplot2)
 
 sims <- rbind(sim_mnl_pref, sim_mnl_wtp, sim_mxl_pref, sim_mxl_wtp)
+sims <- sims[sims$obsID == 1,]
 sims$model <- c(rep("mnl_pref", 4), rep("mnl_wtp", 4),
                 rep("mxl_pref", 4), rep("mxl_wtp", 4))
-sims$alt <- rep(row.names(alts), 4)
+sims$alt <- rep(c("dannon", "hiland", "weight", "yoplait"), 4)
 
-ggplot(sims, aes(x = alt, y = share_mean, fill = model)) +
+ggplot(sims, aes(x = alt, y = mean, fill = model)) +
     geom_bar(stat = 'identity', width = 0.7, position = "dodge") +
-    geom_errorbar(aes(ymin = share_low, ymax = share_high),
+    geom_errorbar(aes(ymin = low, ymax = high),
                   width = 0.2, position = position_dodge(width = 0.7)) +
     scale_y_continuous(limits = c(0, 1)) +
     labs(x = 'Alternative', y = 'Expected Share') +